additions, bug fixing
git-svn-id: https://svn.code.sf.net/p/ccpn/code/trunk@8676 8c04fef6-8de3-4ed0-b809-c157a10b9597

+++ python/ccpnmrcore/modules/GuiStripNd.py
@@ -256,7 +256,7 @@ class GuiStripNd(GuiStrip):
   def setPlaneCount(self, n=0, value=1):
 
     planeCount = self.planeToolbar.planeCounts[n]
-    self.changePlaneCount(value/planeCount.value())
+    self.changePlaneCount(int(value/planeCount.value()))
     self.planeToolbar.planeCount.setValue(value)
 
   def _findPeakListView(self, peakList):
